+++ src/foam/blob/FileBlob.java
@@ -1,32 +1,64 @@
+/**
+ * @license
+ * Copyright 2017 The FOAM Authors. All Rights Reserved.
+ * http://www.apache.org/licenses/LICENSE-2.0
+ */
+
 package foam.blob;
 
+import java.io.BufferedInputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.nio.ByteBuffer;
+
 public class FileBlob
-  extends AbstractBlob
+    extends AbstractBlob
 {
-  private java.io.File f_;
-  private java.nio.channels.FileChannel channel_;
+  protected File file_;
+  protected long size_;
 
-  public FileBlob(java.io.File f) {
-    if ( ! f.canRead() ) {
-      throw new RuntimeException("Cannot read given file.");
-    }
+  public FileBlob(File file) {
+    this.file_ = file;
+    this.size_ = file.length();
+  }
 
-    try {
-      f_ = f;
-      channel_ = java.nio.channels.FileChannel.open(f_.toPath());
-    } catch ( java.io.IOException ex ) {
-      throw new RuntimeException(ex);
-    }
+  public File getFile() {
+    return this.file_;
   }
 
   @Override
-  public Buffer read(Buffer buf, long offset) {
-    // TODO
-    throw new RuntimeException("Unimplemented");
+  public Buffer read(Buffer buffer, long offset) {
+    try {
+      int outOffset = 0;
+
+      long length = Math.min(buffer.getLength(), getSize() - offset);
+      if (length < buffer.getLength()) {
+        buffer = buffer.slice(0, length);
+      }
+
+      ByteBuffer bb = buffer.getData();
+      FileInputStream fileInputStream = new FileInputStream(file_);
+      BufferedInputStream bufferedInputStream = new BufferedInputStream(fileInputStream);
+      bufferedInputStream.skip(offset);
+
+      byte[] buf = new byte[(int) length];
+      while ( outOffset < length ) {
+        int bytesRead = bufferedInputStream.read(buf, outOffset, (int) length);
+        bb.put(buf, outOffset, bytesRead);
+        outOffset += bytesRead;
+      }
+
+      bb.rewind();
+      buffer.setData(bb);
+      return buffer;
+    } catch (Throwable t) {
+      t.printStackTrace();
+      return null;
+    }
   }
 
   @Override
   public long getSize() {
-    return f_.length();
+    return this.size_;
   }
 }
