+++ src/foam/blob/InputStreamBlob.java
@@ -1,3 +1,9 @@
+/**
+ * @license
+ * Copyright 2017 The FOAM Authors. All Rights Reserved.
+ * http://www.apache.org/licenses/LICENSE-2.0
+ */
+
 package foam.blob;
 
 import com.google.common.io.ByteStreams;
