+++ src/foam/nanos/blob/HttpBlobService.java
@@ -1,19 +1,7 @@
 /**
  * @license
- * Copyright 2016 Google Inc. All Rights Reserved.
  * Copyright 2017 The FOAM Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * http://www.apache.org/licenses/LICENSE-2.0
  */
 
 package foam.nanos.blob;
@@ -21,7 +9,7 @@ package foam.nanos.blob;
 import foam.blob.Blob;
 import foam.blob.BlobService;
 import foam.blob.Buffer;
-import foam.blob.FdBlob;
+import foam.blob.FileBlob;
 import foam.core.ContextAware;
 import foam.core.X;
 import foam.nanos.NanoService;
@@ -109,8 +97,8 @@ public class HttpBlobService
     Buffer buffer = new Buffer(BUFFER_SIZE, ByteBuffer.allocate(BUFFER_SIZE));
 
     resp.setStatus(resp.SC_OK);
-    if ( blob instanceof FdBlob ) {
-      File file = ((FdBlob) blob).getFile();
+    if ( blob instanceof FileBlob ) {
+      File file = ((FileBlob) blob).getFile();
       resp.setHeader("Content-Type", Files.probeContentType(Paths.get(file.toURI())));
     } else {
       resp.setHeader("Content-Type", "application/octet-stream");
